third_party/blink/web_tests/platform/generic/http/tests/navigation/anchor-frames-cross-origin-expected.txt



--------
Frame: 'child'
--------
This prevents a cross-origin information leak sometimes know as framesniffing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.documentElement.scrollTop == 0 is true
PASS document.documentElement.scrollLeft == 0 is true
PASS successfullyParsed is true

TEST COMPLETE


--------
Frame: 'grandchild'
--------
This is an anchor point named "anchor1".
